+++ ff_standard_lib/src/strategies/client_features/client_side_brokerage.rs
@@ -14,7 +14,7 @@ use crate::standardized_types::subscriptions::SymbolName;
 use crate::standardized_types::symbol_info::{CommissionInfo, SymbolInfo};
 use crate::strategies::client_features::connection_types::ConnectionType;
 use crate::strategies::client_features::server_connections::{send_request, StrategyRequest};
-use crate::strategies::ledgers::Ledger;
+use crate::strategies::ledgers::ledger::Ledger;
 
 pub(crate) const TIME_OUT: Duration = Duration::from_secs(15);
 impl Brokerage {
